refactor(server): session
